+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/utils/AdMapperTest.kt
@@ -1,0 +1,74 @@
+package com.bitmovin.analytics.bitmovin.player.utils
+
+import com.bitmovin.player.model.advertising.Ad
+import com.bitmovin.player.model.advertising.LinearAd
+import com.bitmovin.player.model.advertising.VastAdData
+import com.bitmovin.player.model.advertising.ima.ImaAdData
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.Before
+import org.junit.Test
+import org.mockito.Mock
+import org.mockito.Mockito.`when`
+import org.mockito.MockitoAnnotations
+
+class AdMapperTest {
+
+    @Mock
+    private lateinit var playerAd: Ad
+
+    @Mock
+    private lateinit var linearAd: LinearAd
+
+    @Mock
+    private lateinit var vastAdData: VastAdData
+
+    @Mock
+    private lateinit var imaAdData: ImaAdData
+
+    private var adMapper = AdMapper()
+
+    @Before
+    fun setup(){
+        MockitoAnnotations.initMocks(this)
+    }
+
+    @Test
+    fun FromPlayerAdWithAdDataNullShouldNotSetAdDataValues() {
+        `when`(playerAd.data).thenReturn(null)
+        var collectorAd = adMapper.fromPlayerAd(playerAd)
+
+        assertThat(collectorAd.bitrate).isNull()
+        assertThat(collectorAd.title).isNull()
+    }
+
+    @Test
+    fun FromPlayerAdWithLinearAdShouldSetLinearAdValues() {
+        val TEN_SENCONDS = 10.0
+        val TEN_SECONDS_IN_MS = TEN_SENCONDS.toLong() * 1000
+
+        `when`(linearAd.duration).thenReturn(TEN_SENCONDS)
+        var collectorAd = adMapper.fromPlayerAd(linearAd)
+
+        assertThat(collectorAd.duration).isEqualTo(TEN_SECONDS_IN_MS)
+    }
+
+    @Test
+    fun FromPlayerAdWithVastAdDataShouldSetVastAdDataValues(){
+        val TITLE = "Title"
+        `when`(playerAd.data).thenReturn(vastAdData)
+        `when`((playerAd.data as VastAdData).adTitle).thenReturn(TITLE)
+        var collectorAd = adMapper.fromPlayerAd(playerAd)
+
+        assertThat(collectorAd.title).isEqualTo(TITLE)
+    }
+
+    @Test
+    fun FromPlayerAdWithImaAdDataShouldSetImaAdDataValues(){
+        val DEALID = "dealid"
+        `when`(playerAd.data).thenReturn(imaAdData)
+        `when`((playerAd.data as ImaAdData).dealId).thenReturn(DEALID)
+        var collectorAd = adMapper.fromPlayerAd(playerAd)
+
+        assertThat(collectorAd.dealId).isEqualTo(DEALID)
+    }
+}