+++ collector/src/main/java/com/bitmovin/analytics/BitmovinAnalytics.java
@@ -5,7 +5,9 @@ import static com.bitmovin.analytics.utils.DataSerializer.serialize;
 import android.content.Context;
 import android.util.Log;
 
+import com.bitmovin.analytics.adapters.AdAdapter;
 import com.bitmovin.analytics.adapters.PlayerAdapter;
+import com.bitmovin.analytics.data.AdEventData;
 import com.bitmovin.analytics.data.ErrorCode;
 import com.bitmovin.analytics.data.EventData;
 import com.bitmovin.analytics.data.IEventDataDispatcher;
@@ -26,7 +28,9 @@ public class BitmovinAnalytics implements StateMachineListener, LicenseCallback 
 
     protected final BitmovinAnalyticsConfig bitmovinAnalyticsConfig;
     protected PlayerAdapter playerAdapter;
+    protected AdAdapter adAdapter;
     protected PlayerStateMachine playerStateMachine;
+    protected BitmovinAdAnalytics adAnalytics;
     protected IEventDataDispatcher eventDataDispatcher;
     protected Context context;
 
@@ -46,6 +50,9 @@ public class BitmovinAnalytics implements StateMachineListener, LicenseCallback 
         this.playerStateMachine = new PlayerStateMachine(this.bitmovinAnalyticsConfig, this);
         this.playerStateMachine.addListener(this);
         this.eventDataDispatcher = new SimpleEventDataDispatcher(this.bitmovinAnalyticsConfig, this.context, this);
+        if(this.bitmovinAnalyticsConfig.getAds()) {
+            this.adAnalytics = new BitmovinAdAnalytics(this);
+        }
     }
 
     /**
@@ -72,10 +79,17 @@ public class BitmovinAnalytics implements StateMachineListener, LicenseCallback 
         this.playerAdapter = adapter;
     }
 
+    protected void attachAd(AdAdapter adapter) {
+        detachAd();
+        this.adAdapter = adapter;
+    }
+
     /**
      * Detach the current player that is being used with Bitmovin Analytics.
      */
     public void detachPlayer() {
+        detachAd();
+
         if (playerAdapter != null) {
             playerAdapter.release();
         }
@@ -84,6 +98,12 @@ public class BitmovinAnalytics implements StateMachineListener, LicenseCallback 
             playerStateMachine.resetStateMachine();
         }
         eventDataDispatcher.disable();
+    }
+
+    private void detachAd() {
+        if(adAdapter != null) {
+            adAdapter.release();
+        }
     }
 
     @Override
@@ -257,6 +277,10 @@ public class BitmovinAnalytics implements StateMachineListener, LicenseCallback 
 
     public void sendEventData(EventData data) {
         this.eventDataDispatcher.add(data);
+    }
+
+    public void sendAdEventData(AdEventData data) {
+        this.eventDataDispatcher.addAd(data);
     }
 
     public long getPosition() {