+++ collector/src/main/java/com/bitmovin/analytics/adapters/AdAdapter.kt
@@ -1,0 +1,10 @@
+package com.bitmovin.analytics.adapters
+
+import com.bitmovin.analytics.data.AdModuleInformation
+
+interface AdAdapter {
+    fun release()
+    val isLinearAdActive: Boolean
+    val moduleInformation: AdModuleInformation
+    val isAutoplayEnabled : Boolean
+}