+++ collector/src/main/java/com/bitmovin/analytics/ads/AdQuartile.kt
@@ -1,0 +1,7 @@
+package com.bitmovin.analytics.ads
+
+enum class AdQuartile {
+    FIRST_QUARTILE,
+    MIDPOINT,
+    THIRD_QUARTILE
+}