+++ collector/src/main/java/com/bitmovin/analytics/ads/AdTagType.kt
@@ -1,0 +1,8 @@
+package com.bitmovin.analytics.ads
+
+enum class AdTagType(name: String) {
+    VAST("vast"),
+    VMAP("vmap"),
+    VPAID("vpaid"),
+    UNKNOWN("unknown")
+}