+++ collector/src/main/java/com/bitmovin/analytics/data/Backend.java
@@ -2,4 +2,5 @@ package com.bitmovin.analytics.data;
 
 public interface Backend {
     void send(EventData eventData);
+    void sendAd(AdEventData eventData);
 }