+++ collector/src/main/java/com/bitmovin/analytics/data/HttpBackend.java
@@ -11,11 +11,14 @@ import com.bitmovin.analytics.utils.HttpClient;
 public class HttpBackend implements Backend {
     private static final String TAG = "BitmovinBackend";
     private HttpClient httpClient;
+    private String analyticsBackendUrl;
+    private String adsAnalyticsBackendUrl;
 
     public HttpBackend(CollectorConfig config, Context context) {
-        String backendUrl = Uri.parse(config.getBackendUrl()).buildUpon().appendEncodedPath("analytics").build().toString();
-        Log.d(TAG, String.format("Initialized Analytics HTTP Backend with %s", backendUrl));
-        this.httpClient = new HttpClient(context, backendUrl);
+        analyticsBackendUrl = Uri.parse(config.getBackendUrl()).buildUpon().appendEncodedPath("analytics").build().toString();
+        adsAnalyticsBackendUrl = Uri.parse(config.getBackendUrl()).buildUpon().appendEncodedPath("analytics/a").build().toString();
+        Log.d(TAG, String.format("Initialized Analytics HTTP Backend with %s", analyticsBackendUrl));
+        this.httpClient = new HttpClient(context);
     }
 
     @Override
@@ -26,7 +29,17 @@ public class HttpBackend implements Backend {
                 eventData.getState(),
                 eventData.getStartupTime(),
                 eventData.getVideoStartupTime(),
-                eventData.getBuffered(), eventData.getAudioLanguage()));
-        this.httpClient.post(DataSerializer.serialize(eventData), null);
+                eventData.getBuffered(), 
+                eventData.getAudioLanguage()));
+        this.httpClient.post(analyticsBackendUrl, DataSerializer.serialize(eventData), null);
+    }
+
+    @Override
+    public void sendAd(AdEventData eventData) {
+        Log.d(TAG, String.format("Sending ad sample: %s (videoImpressionId: %s, adImpressionId: %s)",
+                eventData.getAdImpressionId(),
+                eventData.getVideoImpressionId(),
+                eventData.getAdImpressionId()));
+        this.httpClient.post(adsAnalyticsBackendUrl, DataSerializer.serialize(eventData), null);
     }
 }